reubicado en una carpeta mas especifica y agregado el fetch, el hook useGeneros y los componentes ListadoGeneros e InfoScore
